Refactor service, controller user and handle errors api

+++ src/controllers/User.controller.ts
@@ -2,18 +2,17 @@ import { Request, Response } from 'express';
 import { UserService } from '../services';
 import { validationResult } from 'express-validator';
 import { formatResponse } from '../utils/response.util';
-import { IUserDocument } from '../interfaces/User.interface';
+import { IUserCreate, IUserDocument } from '../interfaces/User.interface';
 import { UserMessages } from '../constants/messages';
+import { HTTP } from '../constants/https';
+import { ApiError } from '../utils/ApiError';
 
 const userService = new UserService();
 
-export const register = async (
-  req: Request,
-  res: Response
-) => {
+export const register = async ( req: Request, res: Response) => {
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
-    res.status(400).json(
+    return res.status(HTTP.BAD_REQUEST).json(
       formatResponse(
         'error',
         errors.array()[0].msg
@@ -22,10 +21,14 @@ export const register = async (
   }
 
   try {
-    const user = await userService.createUser(req.body);
+    const dataUserCreate = {
+      ...req.body,
+      createdBy: req.user?._id
+    } as IUserCreate;
+    const user = await userService.createUser(dataUserCreate, req.user as IUserDocument);
     const token = user.generateAuthToken();
 
-    res.status(201).json(
+    return res.status(HTTP.CREATED).json(
       formatResponse(
         'success',
         UserMessages.REGISTER_SUCCESSFULLY,
@@ -34,7 +37,7 @@ export const register = async (
       )
     )
   } catch (error: any) {
-    res.status(500).json(
+    return res.status(HTTP.INTERNAL_ERROR).json(
       formatResponse(
         'error',
         `Xảy ra lỗi: ${error}`
@@ -43,16 +46,13 @@ export const register = async (
   }
 }
 
-export const login = async(
-  req: Request,
-  res: Response
-) => {
+export const login = async(req: Request,res: Response) => {
   try {
-    const { email, password } = req.body
-    const user = await userService.authenticate(email, password)
-    const token = user.generateAuthToken()
+    const { email, password } = req.body;
+    const user = await userService.authenticate(email, password);
+    const token = user.generateAuthToken();
 
-    res.status(200).json(
+    return res.status(HTTP.OK).json(
       formatResponse(
         'success',
         UserMessages.LOGIN_SUCCESSFULLY,
@@ -63,7 +63,7 @@ export const login = async(
       )
     )
   } catch (error: any) {
-    res.status(500).json(
+    return res.status(500).json(
       formatResponse(
         'error',
         `Xảy ra lỗi: ${error}`
@@ -72,30 +72,22 @@ export const login = async(
   }
 }
 
-export const getCurrentUser = async(
-  req: Request,
-  res: Response
-) => {
+export const getCurrentUser = async(req: Request, res: Response) => {
   try {
     if (!req.user) {
-      res.status(401).json(
-        formatResponse(
-          'error',
-          UserMessages.AUTH_ERROR
-        )
-      )
+      throw new ApiError(HTTP.UNAUTHORIZED, UserMessages.AUTH_ERROR);
     }
   
-    const user = await userService.getUserById(req.params.id.toString())
-    res.status(200).json(
+    const user = await userService.getUserById(req.params.id.toString());
+    return res.status(HTTP.OK).json(
       formatResponse(
         'success',
         UserMessages.GET_CURRENT_USER_SUCCESSFULLY,
         user
       )
-    )
+    );
   } catch (error: any) {
-    res.status(500).json(
+    return res.status(HTTP.INTERNAL_ERROR).json(
       formatResponse(
         'error',
         `Lỗi hệ thống: ${error}`
@@ -104,18 +96,10 @@ export const getCurrentUser = async(
   }
 }
 
-export const updateUser = async(
-  req: Request,
-  res: Response
-) => {
+export const updateUser = async(req: Request, res: Response) => {
   try {
     if (!req.user) {
-      res.status(401).json(
-        formatResponse(
-          'error',
-          UserMessages.AUTH_ERROR
-        )
-      )
+      throw new ApiError(HTTP.UNAUTHORIZED, UserMessages.AUTH_ERROR);
     }
 
     // Chỉ cho phép cập nhật các trường này
@@ -128,21 +112,23 @@ export const updateUser = async(
     );
 
     if (!isValidOperation) {
-      res.status(400).json(
-        formatResponse(
-          'error',
-          UserMessages.ALLOW_FIELDS_NAME_OR_PASSS_UPDATE
-        )
-      )
+      throw new ApiError(HTTP.BAD_REQUEST, UserMessages.ALLOW_FIELDS_NAME_OR_PASSS_UPDATE);
     }
+
+    // Thêm người cập nhật
+    const dataUpdate = {
+      ...req.body,
+      updatedBy: req.user._id
+    } as IUserDocument;
 
     // Cập nhật user từ request.user
     const user = await userService.updateUser(
       req.params.id.toString(),
-      req.body
-    )
+      dataUpdate,
+      req.user as IUserDocument
+    );
 
-    res.status(200).json(
+    return res.status(200).json(
       formatResponse(
         'success',
         UserMessages.UPDATE_SUCCESSFULLY,
@@ -150,7 +136,7 @@ export const updateUser = async(
       )
     )
   } catch (error: any) {
-    res.status(500).json(
+    return res.status(500).json(
       formatResponse(
         'error',
         `Xảy ra lỗi: ${error}`
@@ -159,28 +145,25 @@ export const updateUser = async(
   }
 }
 
-export const deleteUser = async(
-  req: Request,
-  res: Response
-) => {
+export const deleteUser = async(req: Request,res: Response) => {
   try {
+    if (!req.user) {
+      throw new ApiError(HTTP.UNAUTHORIZED, UserMessages.AUTH_ERROR);
+    }
+
     const user = await userService.deleteUser(req.params.id.toString());
     if (!user) {
-      res.status(404).json(
-        formatResponse(
-          'error',
-          UserMessages.USER_NOT_FOUND
-        )
-      )
+      throw new ApiError(HTTP.NOT_FOUND, UserMessages.USER_NOT_FOUND);
     }
-    res.status(200).json(
+
+    return res.status(HTTP.OK).json(
       formatResponse(
         'success',
         UserMessages.USER_UPDATE_STATUS_SUCCESSFULLY
       )
     )
   } catch (error: any) {
-    res.status(500).json(
+    return res.status(HTTP.INTERNAL_ERROR).json(
       formatResponse(
         'error',
         `Xảy ra lỗi: ${error}`
@@ -189,30 +172,29 @@ export const deleteUser = async(
   }
 }
 
-export const getUsers = async(
-  req: Request,
-  res: Response
-) => {  
+export const getUsers = async(req: Request,res: Response) => {  
   try {
     const { 
+      q,
       page = 1, 
       limit = 10, 
       sortBy = 'createdAt', 
       sortOrder = 'desc' 
     } = req.query as {
+      q: string;
       page?: string;
       limit?: string;
       sortBy?: 'createdAt' | 'name' | 'email';
       sortOrder?: 'asc' | 'desc';
     }
-    const result = await userService.getPaginatedUsers({
+    const result = await userService.getPaginatedUsers(q, {
       page: parseInt(page as string),
       limit: parseInt(limit as string),
       sortBy,
       sortOrder
-    })
+    });
 
-    res.status(200).json(
+    return res.status(HTTP.OK).json(
       formatResponse(
         'success',
         UserMessages.GET_USERS_SUCCESSFULLY,
@@ -220,73 +202,60 @@ export const getUsers = async(
         undefined,
         result.pagination
       )
-    )
+    );
   } catch (error: any) {
-    res.status(500).json(
+    return res.status(HTTP.INTERNAL_ERROR).json(
       formatResponse(
         'error',
         `Xảy ra lỗi: ${error}`
       )
-    )
+    );
   }
 }
 
-export const searchUsers = async(
-  req: Request,
-  res: Response
-) => {
+// export const searchUsers = async(
+//   req: Request,
+//   res: Response
+// ) => {
+//   try {
+//     const { q, page = 1, limit = 10 } = req.query as {
+//       q: string;
+//       page?: string;
+//       limit?: string;
+//     }
+
+//     const result = await userService.searchUsers(q, {
+//       page: parseInt(page as string),
+//       limit: parseInt(limit as string)
+//     })
+
+//     res.status(200).json(
+//       formatResponse(
+//         'success',
+//         UserMessages.SEARCH_USER_SUCCESSFULLY,
+//         result.data,
+//         undefined,
+//         result.pagination
+//       )
+//     )
+//   } catch (error: any) {
+//     res.status(500).json(
+//       formatResponse(
+//         'error',
+//         `Xảy ra lỗi: ${error}`
+//       )
+//     )
+//   }
+// }
+
+export const updateAvatar = async(req: Request,res: Response) => {
   try {
-    const { q, page = 1, limit = 10 } = req.query as {
-      q: string;
-      page?: string;
-      limit?: string;
-    }
-
-    const result = await userService.searchUsers(q, {
-      page: parseInt(page as string),
-      limit: parseInt(limit as string)
-    })
-
-    res.status(200).json(
-      formatResponse(
-        'success',
-        UserMessages.SEARCH_USER_SUCCESSFULLY,
-        result.data,
-        undefined,
-        result.pagination
-      )
-    )
-  } catch (error: any) {
-    res.status(500).json(
-      formatResponse(
-        'error',
-        `Xảy ra lỗi: ${error}`
-      )
-    )
-  }
-}
-
-export const updateAvatar = async(
-  req: Request,
-  res: Response
-) => {
-  try {
-    if(!req.user) {
-      return res.status(401).json(
-        formatResponse(
-          'error',
-          'Forbidden'
-        )
-      )
+    if (!req.user) {
+      throw new ApiError(HTTP.UNAUTHORIZED, UserMessages.AUTH_ERROR);
     }
 
     if(!req.file) {
-      return res.status(500).json(
-        formatResponse(
-          'error',
-          UserMessages.USER_CHOOSE_AVT
-        )
-      )
+      throw new ApiError(HTTP.INTERNAL_ERROR, UserMessages.USER_CHOOSE_AVT);
     }
 
     const user = await userService.updateUser(
@@ -296,7 +265,7 @@ export const updateAvatar = async(
       } as IUserDocument
     )
 
-    res.status(200).json(
+    return res.status(HTTP.OK).json(
       formatResponse(
         'success',
         UserMessages.USER_UPDATE_AVT_SUCCESSFULLY,
@@ -304,7 +273,7 @@ export const updateAvatar = async(
       )
     )
   } catch (error: any){
-    res.status(500).json(
+    return res.status(HTTP.INTERNAL_ERROR).json(
       formatResponse(
         'error',
         `Xảy ra lỗi: ${error}`
@@ -313,43 +282,29 @@ export const updateAvatar = async(
   }
 }
 
-export const changePassword = async(
-  req: Request,
-  res: Response
-) => {
+export const changePassword = async(req: Request,res: Response) => {
   try {
-    if(!req.user) {
-      return res.status(401).json(
-        formatResponse(
-          'error',
-          'Forbidden'
-        )
-      )
+    if (!req.user) {
+      throw new ApiError(HTTP.UNAUTHORIZED, UserMessages.AUTH_ERROR);
     }
 
-    const { oldPassword, newPassword } = req.body
-    const isMatch = await req.user.comparePassword(oldPassword)
+    const { oldPassword, newPassword } = req.body;
+    const isMatch = await req.user.comparePassword(oldPassword);
 
     if(!isMatch) {
-      return res.status(500).json(
-        formatResponse(
-          'error',
-          UserMessages.PASSWORD_INCORRECT
-        )
-      )
+      throw new ApiError(HTTP.INTERNAL_ERROR, UserMessages.PASSWORD_INCORRECT);
     }
 
-    const user = await userService.changePassword(req.user.id, newPassword)
-    res.status(200).json(
+    const user = await userService.changePassword(req.user.id, newPassword);
+    return res.status(HTTP.OK).json(
       formatResponse(
         'success',
         UserMessages.UPDATE_PASSWORD_SUCCESSFULLY,
         user
       )
-    )
-
+    );
   } catch (error: any) {
-    res.status(500).json(
+    return res.status(HTTP.INTERNAL_ERROR).json(
       formatResponse(
         'error',
         `Xảy ra lỗi: ${error}`
